backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
pydantic[email]
bcrypt
pyjwt
python-multipart
requests
httpx
psycopg[binary]
aiofiles
sendgrid